requirements.txt
django=4.2.5
requests=2.31.0
beautifulsoup4=4.12.2
psycopg2-binary=3.1.11
flower=2.0.1
websocket-client=1.6.3
django-redis=5.3.0
celery=5.3.4
